Add new property to force screenshot with same size as video source
This new property is useful when you request a larger video source
via e.g. 'videoConstraints={{ width: 1080}}', but display it in a
smaller preview place. When enabled, the screenshot has the full size
of the source video stream, which also means that no resizing is
necessary when drawing it to the canvas.

Default behavior does not change.